sui_indexer_alt_framework::models::cp_sequence_numbers
Function epoch_interval

pub async fn epoch_interval(
    conn: &mut Connection<'_>,
    cps: Range<u64>,
) -> Result<Range<u64>>

Returns the epochs of the given checkpoint range. start is the epoch of the first checkpoint and end is the epoch of the last checkpoint.