Module sui_indexer_alt_framework::pipeline

Modules

concurrent

sequential

Structs

CommitterConfig

Traits

Processor

Implementors of this trait are responsible for transforming checkpoint into rows for their table. The FANOUT associated value controls how many concurrent workers will be used to process checkpoint information.