sui_indexer_alt_framework::postgres::schema
Module watermarks

Modules

columns

Contains all of the columns of this table

dsl

Re-exports all of the columns of this table, as well as the table struct renamed to the module name. This is meant to be glob imported for functions which only deal with one table.

Structs

checkpoint_hi_inclusive

epoch_hi_inclusive

pipeline

pruner_hi

pruner_timestamp

reader_lo

star

Represents table_name.*, which is sometimes needed for efficient count queries. It cannot be used in place of all_columns, and has a SqlType of () to prevent it being used that way

table

The actual table struct

timestamp_ms_hi_inclusive

tx_hi

Constants

all_columns

A tuple of all of the columns on this table

Type Aliases

BoxedQuery

Helper type for representing a boxed query from this table

SqlType

The SQL type of all of the columns on this table