sui_indexer_alt_framework::postgres

Module store

Traits

Store

A storage-agnostic interface that provides database connections for both watermark management and arbitrary writes. The indexer framework accepts this Store implementation to manage watermarks operations through its associated Connection type. This store is also passed to the pipeline handlers to perform arbitrary writes to the store.