sui_indexer_alt_framework::postgres
Module temp

Structs

LocalDatabase

Local instance of a postgres server.

TempDb

A temporary, local postgres database

Functions

get_available_port

Return an ephemeral, available port. On unix systems, the port returned will be in the TIME_WAIT state ensuring that the OS won’t hand out this port for some grace period. Callers should be able to bind to this port given they use SO_REUSEADDR.