sui_indexer_alt_framework

Module cluster

Structs

Args

Bundle of arguments for setting up an indexer cluster (an Indexer and its associated Metrics service). This struct is offered as a convenience for the common case of parsing command-line arguments for a binary running a standalone indexer and its metrics service.

IndexerCluster

An opinionated IndexerCluster that spins up an Indexer implementation using Postgres as its store, along with a MetricsService and a tracing subscriber (outputting to stderr) to provide observability. It is a useful starting point for an indexer binary.

IndexerClusterBuilder

Builder for creating an IndexerCluster with a fluent API