Trait sui_indexer_alt_framework::pipeline::Processor

pub trait Processor {
    type Value: Send + Sync + 'static;

    const NAME: &'static str;
    const FANOUT: usize = 10usize;

    // Required method
    fn process(
        &self,
        checkpoint: &Arc<CheckpointData>,
    ) -> Result<Vec<Self::Value>>;
}

Implementors of this trait are responsible for transforming checkpoint into rows for their table. The FANOUT associated value controls how many concurrent workers will be used to process checkpoint information.

Required Associated Types

type Value: Send + Sync + 'static

The type of value being inserted by the handler.

Required Associated Constants

const NAME: &'static str

Used to identify the pipeline in logs and metrics.

Provided Associated Constants

const FANOUT: usize = 10usize

How much concurrency to use when processing checkpoint data.

Required Methods

fn process(&self, checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>>

The processing logic for turning a checkpoint into rows of the table.

Object Safety

This trait is not object safe.