sui_indexer_alt_framework
Module postgres

Modules

query

schema

store

temp

Macros

sql

The sql! macro is used to construct a diesel::SqlLiteral<T> using a format string to describe the SQL snippet with the following syntax:

Structs

Connection

Wrapper struct over the remote PooledConnection type for dealing with the Store trait.

Db

DbArgs

Constants

MIGRATIONS

Traits

FieldCount

Trait that provides a constant indicating the number of fields in a struct.

Functions

merge_migrations

Returns new migrations derived from the combination of provided migrations and migrations defined in this crate.

reset_database

Drop all tables, and re-run migrations if supplied.

Derive Macros

FieldCount